backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
aiosmtplib==3.0.2
email-validator==2.2.0
pydantic[email]==2.10.4
passlib[bcrypt]==1.7.4
sqladmin==0.20.1
itsdangerous==2.2.0
Babel==2.18.0
